requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
cffi==1.17.1
cryptography==45.0.4
pycparser==2.22
sqlparse==0.5.3
django-allauth==65.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
requests==2.32.4
urllib3==2.5.0
pillow>=10.2.0



